+++ packages/shared/src/enum.ts
@@ -82,6 +82,7 @@ export const APP_DEFAULT_CONFIG: AppConfig = {
   /** 允许自定义可执行文件地址 */
   customExecPath: false,
   requestInfoForRecord: true,
+  cutPageInNewWindow: false,
   bilibiliUser: {},
   tool: {
     home: {
@@ -247,6 +248,8 @@ export const APP_DEFAULT_CONFIG: AppConfig = {
     quality: "highest",
     line: undefined,
     checkInterval: 60,
+    maxThreadCount: 3,
+    waitTime: 0,
     disableProvideCommentsWhenRecording: false,
     segment: 90,
     saveGiftDanma: false,