+++ packages/liveManager/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 1.8.0
 
 - 标题黑名单增加额外状态
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`